# spec/swagger.yaml
definitions:
...
  display.SessionInfo:
    properties:
      codec:
        type: string
      created_at:
        type: string
      id:
        type: string
      state:
        type: string
    type: object
...
  handlers.ErrorResponse:
    properties:
      message:
        type: string
    type: object
...
  handlers.createSessionRequest:
    properties:
      channel_type:
        type: string
      metadata:
        additionalProperties: {}
        type: object
      title:
        type: string
      type:
        type: string
    type: object
...
  handlers.displaySessionListResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/display.SessionInfo'
        type: array
    type: object
...
  handlers.updateSessionRequest:
    properties:
      metadata:
        additionalProperties: {}
        type: object
      title:
        type: string
      type:
        type: string
    type: object
...
  session.Session:
    properties:
      bot_id:
        type: string
      channel_type:
        type: string
      created_at:
        type: string
      id:
        type: string
      metadata:
        additionalProperties: {}
        type: object
      parent_session_id:
        type: string
      route_conversation_type:
        type: string
      route_id:
        type: string
      route_metadata:
        additionalProperties: {}
        type: object
      title:
        type: string
      type:
        type: string
      updated_at:
        type: string
    type: object
...
info:
  contact: {}
  title: Memoh API
  version: 1.0.0
paths:
...
  /bots/{bot_id}/container/display/sessions:
    get:
      parameters:
      - description: Bot ID
        in: path
        name: bot_id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.displaySessionListResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: List active workspace display WebRTC sessions
      tags:
      - containerd
  /bots/{bot_id}/container/display/sessions/{session_id}:
    delete:
      parameters:
      - description: Bot ID
        in: path
        name: bot_id
        required: true
        type: string
      - description: Display session ID
        in: path
        name: session_id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Close a workspace display WebRTC session
      tags:
      - containerd
...
  /bots/{bot_id}/sessions:
    get:
      parameters:
      - description: Bot ID
        in: path
        name: bot_id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              items:
                $ref: '#/definitions/session.Session'
              type: array
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: List bot sessions
      tags:
      - sessions
    post:
      parameters:
      - description: Bot ID
        in: path
        name: bot_id
        required: true
        type: string
      - description: Session data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.createSessionRequest'
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/session.Session'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Create a new chat session
      tags:
      - sessions
  /bots/{bot_id}/sessions/{session_id}:
    delete:
      parameters:
      - description: Bot ID
        in: path
        name: bot_id
        required: true
        type: string
      - description: Session ID
        in: path
        name: session_id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Soft-delete a session
      tags:
      - sessions
    get:
      parameters:
      - description: Bot ID
        in: path
        name: bot_id
        required: true
        type: string
      - description: Session ID
        in: path
        name: session_id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/session.Session'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Get a session by ID
      tags:
      - sessions
    patch:
      parameters:
      - description: Bot ID
        in: path
        name: bot_id
        required: true
        type: string
      - description: Session ID
        in: path
        name: session_id
        required: true
        type: string
      - description: Fields to update
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.updateSessionRequest'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/session.Session'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Update a session
      tags:
      - sessions
...
swagger: "2.0"